+++ src/components/LettersNav.jsx
@@ -165,9 +165,9 @@ function LettersNav() {
           })}
         </StUl>
       </StNav>
-      <LetterForm onSubmitLetter={onSubmitLetter} />
+      <LetterForm />
       {/* Item 컴포넌트부분 */}
-      <Item name={name} />
+      <Item />
     </>
   );
 }
