+++ src/shared/Router.js
@@ -1,6 +1,5 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import LettersLayout from "components/LettersLayout";
-import Item from "components/Item";
 import Detail from "Pages/Detail";
 
 const Router = () => {
@@ -8,10 +7,7 @@ const Router = () => {
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<LettersLayout />} />
-        <Route path="item" element={<Item />} />
         <Route path="navi/:id" element={<Detail />} />
-        <Route path="detail" element={<Detail />} />
-        <Route path="item" element={<Item />} />
       </Routes>
     </BrowserRouter>
   );
